compiler: Move Object hyerarchy to separate module
